test(e2e): rename tiers — smoke (always-on) and full (on demand)
- specs move to src/tests/smoke/ and src/tests/full/
- full = smoke + heavy specs; e2e:loop --full / e2e --full, E2E_FULL gate
- CI and default runs unchanged: smoke only
