+++ packages/e2e/playwright/src/lib/auth.ts
@@ -38,7 +38,7 @@ export async function getAccessToken(
 /** Accept the current Terms of Use for the client's user. The terms gate
  *  412s every non-terms API call until the user accepts, and only the UI
  *  login flow (01-auth) does it implicitly — self-contained API specs
- *  (extended suites) must call this before anything else. */
+ *  (the full-suite tier) must call this before anything else. */
 export async function acceptTerms(api: ApiClient): Promise<void> {
   const current = await api.terms.current.query();
   await api.terms.accept.mutate({ version: current.version });
