+++ packages/e2e/playwright/src/tests/smoke/04-messages.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from "@playwright/test";
 
-import { baseUrl } from "../config.js";
+import { baseUrl } from "../../config.js";
 import {
   agentCardStatus,
   chatInput,
@@ -10,10 +10,10 @@ import {
   setMockAgentReply,
   setMockReplyWithMidTurnUserPrompt,
   waitForAgentRunning,
-} from "../lib/agents.js";
-import { createApiClient } from "../lib/api-client.js";
-import { getAccessToken } from "../lib/auth.js";
-import { agentName } from "../lib/fixtures.js";
+} from "../../lib/agents.js";
+import { createApiClient } from "../../lib/api-client.js";
+import { getAccessToken } from "../../lib/auth.js";
+import { agentName } from "../../lib/fixtures.js";
 
 const scriptedReply = "scripted-reply-from-e2e";
 const userPrompt = "hello-from-playwright";
