+++ packages/e2e/playwright/tasks.toml
@@ -13,15 +13,16 @@ fi
 '''
 
 ["e2e-playwright:run"]
+description = "Run the smoke suite (src/tests/smoke/) — the always-on pipeline tier"
 dir = "{{config_root}}/packages/e2e/playwright"
 depends = ["common:setup:pnpm", "e2e-playwright:install-browsers"]
 run = "pnpm exec playwright test"
 
-["e2e-playwright:run-extended"]
-description = "Run ONLY the extended suites (slow, manually-triggered specs under src/tests/extended/)"
+["e2e-playwright:run-full"]
+description = "Run the FULL suite: smoke plus the slow on-demand specs under src/tests/full/"
 dir = "{{config_root}}/packages/e2e/playwright"
 depends = ["common:setup:pnpm", "e2e-playwright:install-browsers"]
-run = "E2E_EXTENDED=1 pnpm exec playwright test --project '*-extended'"
+run = "E2E_FULL=1 pnpm exec playwright test"
 
 ["e2e-playwright:run-headed"]
 description = "Run Playwright with a visible browser window"
